+++ gasapi.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.4.33110.190
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "gasapi", "gasapi.csproj", "{2B6CC72F-B91C-4596-A6AF-6D247CCE2810}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GasApi", "GasApi.csproj", "{2B6CC72F-B91C-4596-A6AF-6D247CCE2810}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "dynamicheating", "..\vsmodtemplate-1.5.7\mods\dynamicheat\dynamicheating.csproj", "{BA28E816-C9D9-429A-A064-10BF89885C8F}"
 EndProject
