+++ src/Blocks/BlockSolid.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using Vintagestory.API.Common;
+
+namespace GasApi.Blocks
+{
+    internal class BlockSolid : Block
+    {
+    }
+}
